metrics_v2/handler/requirements.txt
absl-py==0.11.0
google-auth==1.23.0
google-cloud-bigquery==2.5.0
functions-framework==2.1.0
google-cloud-secret-manager==2.1.0
jinja2==2.11.2
kubernetes==12.0.1
numpy==1.18.5
protobuf==3.14.0
pytz==2020.4
sendgrid==6.4.8
tensorboard==2.4.0
tensorflow-cpu==2.4.0
